backend/domain-service/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytest==6.2.5
dnspython==2.1.0
sqlalchemy_utils